Move FeedbackApplication-class to package mops
