+++ src/main/java/mops/FeedbackApplication.java
@@ -1,4 +1,4 @@
-package mops.feedback;
+package mops;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
